+++ src/stories/components/atoms/Select/SelectPerson/index.stories.ts
@@ -1,5 +1,3 @@
-import { action } from '@storybook/addon-actions';
-
 import type { Meta, StoryObj } from '@storybook/react';
 
 import { SelectPerson } from '.';
@@ -20,7 +18,6 @@ type Story = StoryObj<typeof meta>;
 
 export const SelectPersonInit: Story = {
   args: {
-    selectHandler: action('click'),
     data: data['0001'].addressList,
   },
 };
