+++ src/stories/components/atoms/Select/SelectPerson/index.tsx
@@ -1,34 +1,13 @@
-import { InputLabel, SelectChangeEvent } from '@mui/material';
-import { FC, useContext } from 'react';
+import { InputLabel } from '@mui/material';
+import { FC } from 'react';
 
+import { useSelectPersonEvent } from './hooks';
 import { FormControl } from '../../FormControl/Basic';
 import { SelectProps } from '../Base';
 import { Select } from '../Basic';
 
-import {
-  InitChangeEventStateContext,
-  ResetSendStateContext,
-  SelectedPersonIdContext,
-} from '@/stories/common/context';
-// import { handleChange } from './functions';
-
 export const SelectPerson: FC<SelectProps> = ({ data }: SelectProps) => {
-  const { initialChangeOccurred, setInitialChangeOccurred } =
-    useContext(InitChangeEventStateContext) ?? {};
-  const { sendState, dispatch } = useContext(ResetSendStateContext) ?? {};
-  const selectedPersonId = useContext(SelectedPersonIdContext);
-  const handleChange = (event: SelectChangeEvent<unknown>) => {
-    if (selectedPersonId) {
-      selectedPersonId(event.target.value as string);
-    }
-    if (!initialChangeOccurred && setInitialChangeOccurred) {
-      setInitialChangeOccurred(true);
-    }
-    if (sendState === 2 && dispatch) {
-      dispatch('INIT');
-    }
-  };
-
+  const { handleChange } = useSelectPersonEvent();
   return (
     <FormControl>
       <InputLabel id='select-label'>To</InputLabel>
